backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Database
tortoise-orm>=0.20.0
asyncpg>=0.29.0

# AI and NLP
spacy>=3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

# Caching and messaging
redis>=5.0.1
aioredis>=2.0.1

# Monitoring and logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7
structlog>=24.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
httpx>=0.25.2

# Security
python-keycloak>=3.3.0
cryptography>=41.0.5 